pyproject.toml
[tool.poetry]
package-mode = false
readme = "README.md"
homepage = 'http://0.0.0.0:8000/'
repository = 'http://0.0.0.0:8000/'

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
qrcode = {extras = ["pil"], version = "^7.4.2"}
lxml = "^5.3.0"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
django-allauth = "^65.0.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
